Enter raw mode
It's not quite working properly yet. The program instantly exits when it
is run. In the loop when Stdin is read, 0 bytes are read immediately at
runtime and the program exits. This doesn't happen if the program has
raw mode removed.

Altering the TERMIOS ECHO is done on Stdin so that could be why no bytes
are read from Stdin later. Or using `os.Stdin.Read()` might not be the
correct function to use.