Turn off output processing: automatic carriage return insertion
When a newline is entered the terminal output automatically adds a
carriage return. This is a holdover from teletype terminals to return
the writing head to the far left column.

OPOST turns off this processing but also means we have to manually add
the carriage return in our print statements.