Disable miscellaneous flags
These flags likely don't have any observable effect because they are
already disabled or don't apply to modern terminal emulators.

At some point switching them off has been considered a part of
'raw mode', so we disable them.

When BRKINT is turned on, a break condition will cause a SIGINT signal
to be sent to the program, like pressing Ctrl-C.

INPCK enables parity checking, which doesn’t seem to apply to modern
terminal emulators.

ISTRIP causes the 8th bit of each input byte to be stripped, meaning it
will set it to 0. This is probably already turned off.

CS8 is not a flag, it is a bit mask with multiple bits, which we set
using the bitwise-OR (|) operator unlike all the flags we are turning
off. It sets the character size (CS) to 8 bits per byte.