Skip adding timeouts and error handling steps
Unlike the online tutorial I've skipped setting a timeout with
VMIN and VTIME in the Cc array of the termios. I'm not sure how it
is supposed to help the program now or later. With the timeout
enabled the program quits after a fraction of a second. This is
correct behavior I just don't know how it benefits me.

I've also skipped the error handling step (the `die` function) from
the tutorial. Go already encourages pretty stringient error
handling ever step of the way with these kind of low level tasks so
adding a custom function just seems redundant.

I've also skipped arranging the code into the same sections as the
tutorial.

See steps 16, 17, 18, and 19:
https://viewsourcecode.org/snaptoken/kilo/02.enteringRawMode.html#a-timeout-for-read